refractored v2 robots into modular_robots_v2
